Conflicts:
	anki/consts.py
	anki/decks.py
	anki/importing/noteimp.py
	anki/media.py
	anki/stats.py
	anki/sync.py
	anki/utils.py
	aqt/addons.py
	aqt/deckconf.py
	aqt/dyndeckconf.py
	aqt/importing.py
	aqt/main.py
	aqt/utils.py

+++ aqt/main.py
@@ -305,14 +305,13 @@ attempting to import."""))
             self.col = Collection(cpath, log=True)
         except anki.db.Error:
             # warn user
-            showWarning("""\
-Your collection is corrupt. Please see the manual for \
-how to restore from a backup.""")
-            # move it out of the way so the profile can be used again
-            newpath = cpath+str(intTime())
-            os.rename(cpath, newpath)
-            # then close
-            sys.exit(1)
+            showWarning(_("""\
+Your collection is corrupt. Please create a new profile, then \
+see the manual for how to restore from an automatic backup.
+
+Debug info:
+""")+traceback.format_exc())
+            self.unloadProfile()
         except Exception as e:
             # the custom exception handler won't catch this if we immediately
             # unload, so we have to manually handle it
@@ -786,8 +785,8 @@ title="%s">%s</button>''' % (
     def onImport(self):
         aqt.importing.onImport(self)
 
-    def onExport(self):
-        aqt.exporting.ExportDialog(self)
+    def onExport(self, did=None):
+        aqt.exporting.ExportDialog(self, did=did)
 
     # Cramming
     ##########################################################################
@@ -991,7 +990,9 @@ will be lost. Continue?"""))
         diag.connect(box, SIGNAL("rejected()"), diag, SLOT("reject()"))
         diag.setMinimumHeight(400)
         diag.setMinimumWidth(500)
+        restoreGeom(diag, "checkmediadb")
         diag.exec_()
+        saveGeom(diag, "checkmediadb")
 
     def deleteUnused(self, unused, diag):
         if not askUser(_("Delete unused media?")):
@@ -999,7 +1000,8 @@ will be lost. Continue?"""))
         mdir = self.col.media.dir()
         for f in unused:
             path = os.path.join(mdir, f)
-            send2trash(path)
+            if os.path.exists(path):
+                send2trash(path)
         tooltip(_("Deleted."))
         diag.close()
 
@@ -1021,11 +1023,13 @@ will be lost. Continue?"""))
         self.progress.finish()
         part1 = ngettext("%d card", "%d cards", len(cids)) % len(cids)
         part1 = _("%s to delete:") % part1
-        diag, box = showText(part1 + "\n\n" + report, run=False)
+        diag, box = showText(part1 + "\n\n" + report, run=False,
+                geomKey="emptyCards")
         box.addButton(_("Delete Cards"), QDialogButtonBox.AcceptRole)
         box.button(QDialogButtonBox.Close).setDefault(True)
 
         def onDelete():
+            saveGeom(diag, "emptyCards")
             QDialog.accept(diag)
             self.checkpoint(_("Delete Empty"))
             self.col.remCards(cids)
